aai-590-capstone-banana_enterprise_system/requirements.txt

fastapi
uvicorn
requests
langgraph
faiss-cpu
transformers
torch
pydantic
numpy
